requirements/local.txt
-r base.txt
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.12.0
pytest==7.4.0
pytest-django==4.5.2
